add button state

+++ controller/listController.js
@@ -6,6 +6,7 @@ app.controller('listController', ['$scope', 'localStorageService', 'blihService'
     function ($scope, localStorageService, blihService, usSpinnerService, $state) {
         const user = localStorageService.get('user')
         $scope.repositoryList = user.repositoryList
+        $scope.isInDelete = false
         const deleteModal = $('#deleteModal')
         const deleteSpinner = 'deleteSpinner'
         const wesh = msg => console.log(msg)
@@ -28,9 +29,10 @@ app.controller('listController', ['$scope', 'localStorageService', 'blihService'
             $scope.deleteRepoName = ''
             deleteModal.modal('hide')
         }
-
-        //TODO(carlendev) add error gestion
+        
         $scope.deleteRepoEvent = () => {
+            if ($scope.isInDelete) return
+            $scope.isInDelete = true
             $scope.startSpin(deleteSpinner)
             blihService.deleteRepo(user.userName, user.token, $scope.deleteRepoName)
                 .then(res => {
@@ -41,11 +43,13 @@ app.controller('listController', ['$scope', 'localStorageService', 'blihService'
                         $scope.$apply()
                         $scope.stopSpin(deleteSpinner)
                         deleteModal.modal('hide')
+                        $scope.isInDelete = false
                     })
                 })
                 .catch(err => {
                     wesh('delete err')
                     wesh(err)
+                    $scope.isInDelete = false
                 })
         }
 }])
